feat: final organization of src
