+++ examples/ex_2_time_evolution.py
@@ -19,7 +19,7 @@ from pprint import pprint
 from benchq.algorithms.time_evolution import qsp_time_evolution_algorithm
 from benchq.quantum_hardware_modeling import BASIC_SC_ARCHITECTURE_MODEL
 from benchq.problem_ingestion import get_vlasov_hamiltonian
-from benchq.problem_ingestion.hamiltonians.hiesenburg import (
+from benchq.problem_ingestion.hamiltonians.hiesenberg_hamiltonians import (
     generate_1d_heisenberg_hamiltonian,
 )
 from benchq.resource_estimators.graph_estimators import (
