+++ src/benchq/problem_embeddings/block_encodings/offset_tridiagonal_utils.py
