+++ src/benchq/problem_embeddings/time_marching/compression_gadget.py
@@ -2,7 +2,7 @@ import math
 
 from orquestra.quantum.circuits import Circuit
 
-from benchq.problem_ingestion.block_encodings.utils.offset_trdiagonal_utils import (
+from benchq.problem_embeddings.block_encodings.utils.offset_tridiagonal_utils import (
     controlled_clock,
 )
