+++ src/benchq/problem_ingestion/__init__.py
@@ -5,8 +5,8 @@ from .hamiltonians.hamiltonian_from_file import (
     get_all_hamiltonians_in_folder,
     get_hamiltonian_from_file,
 )
-from .hamiltonians.jordan_wigner import (
+from .hamiltonians.jordan_wigner_hamiltonian import (
     generate_fermi_hubbard_jw_qubit_hamiltonian,
     generate_jw_qubit_hamiltonian_from_mol_data,
 )
-from .hamiltonians.vlasov import get_vlasov_hamiltonian
+from .hamiltonians.vlasov_hamiltonians import get_vlasov_hamiltonian
