+++ src/benchq/problem_ingestion/hamiltonians/_compute_lambda.py
