+++ src/benchq/problem_ingestion/hamiltonians/ising_hamiltonians.py
