+++ tests/benchq/problem_embeddings/block_encodings/test_offset_tridiagonal.py
