+++ tests/benchq/problem_embeddings/test_openfermion_re.py
@@ -10,8 +10,8 @@ from benchq.quantum_hardware_modeling import BasicArchitectureModel
 from benchq.resource_estimators.footprint_estimators.openfermion_estimator import (
     footprint_estimator,
 )
-from benchq.resource_estimators.footprint_estimators.openfermion_re import (
-    get_double_factorized_block_encoding_info,
+from benchq.problem_embeddings._qpe import (
+    get_double_factorized_block_encoding,
     get_double_factorized_qpe_toffoli_and_qubit_cost,
     get_single_factorized_qpe_toffoli_and_qubit_cost,
 )
@@ -77,7 +77,7 @@ def test_df_block_encoding_logical_qubit_count_is_larger_than_number_of_spin_orb
     mean_field_object = instance.get_active_space_meanfield_object()
     h1, eri_full, _, _, _ = pyscf_to_cas(mean_field_object)
 
-    (num_toffoli, num_qubits, lam) = get_double_factorized_block_encoding_info(
+    (num_toffoli, num_qubits, lam) = get_double_factorized_block_encoding(
         h1, eri_full, 1e-6
     )
     assert num_qubits > 2 * eri_full.shape[0]
@@ -102,7 +102,7 @@ def test_df_block_encoding_lambda_scales_with_hamiltonian(
     threshold = 1e-6
     scale_factor = 10
 
-    (num_toffoli, num_qubits, lam) = get_double_factorized_block_encoding_info(
+    (num_toffoli, num_qubits, lam) = get_double_factorized_block_encoding(
         h1, eri_full, threshold
     )
 
@@ -110,7 +110,7 @@ def test_df_block_encoding_lambda_scales_with_hamiltonian(
         scaled_num_toffoli,
         scaled_num_qubits,
         scaled_lam,
-    ) = get_double_factorized_block_encoding_info(
+    ) = get_double_factorized_block_encoding(
         scale_factor * h1, scale_factor * eri_full, scale_factor * threshold
     )
 
@@ -144,7 +144,7 @@ def test_double_factorized_qpe_raises_exception_for_invalid_eri():
 def test_double_factorized_block_encoding_raises_exception_for_invalid_eri():
     h1, eri_full = _get_asymmetric_hamiltonian()
     with pytest.raises(ValueError):
-        get_double_factorized_block_encoding_info(h1, eri_full, 1e-6)
+        get_double_factorized_block_encoding(h1, eri_full, 1e-6)
 
 
 def test_physical_qubits_larger_than_logical_qubits():
